mzeebikes/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.8.0
python-decouple==3.8
sqlparse==0.5.0
